+++ js/data.js
@@ -1,0 +1,35 @@
+
+const PICTURE_DESCRIPTIONS = [
+  'Природа',
+  'Закат',
+  'Городской сюжет',
+  'Осеннее настроение',
+  'Котики',
+  'Животные в зоопарке',
+  'В кафе', 'Вечер с любимыми',
+  'Роза ветров',
+  'Настоящие чувства',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Алексей',
+  'Кекс',
+  'Гарри',
+  'Рональд',
+  'Стив',
+  'Леонардо',
+  'Форест',
+  'Ричард',
+  'Дональд',
+];
+
+export {NAMES, MESSAGES, PICTURE_DESCRIPTIONS};
