+++ js/utils.js
@@ -1,0 +1,3 @@
+const checkTextLength = (text, maxLength) => text.length <= maxLength;
+
+export {checkTextLength};
